- Fix spelling errors - Update component renames - Remove logic to fix bid cycle name
