+++ src/Components/BidCycleList/BidCycleCard/BidCycleCard.test.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import toJSON from 'enzyme-to-json';
-import BidCycle from './BidCycle';
+import BidCycleCard from './BidCycleCard';
 import mock from '../../../__mocks__/bidCycles';
 
-describe('BidCycleComponent', () => {
+describe('BidCycleCard Component', () => {
   const props = {
     props: mock[0],
   };
 
   it('is defined', () => {
-    const wrapper = shallow(<BidCycle {...props} />);
+    const wrapper = shallow(<BidCycleCard {...props} />);
     expect(wrapper).toBeDefined();
   });
 
   it('matches snapshot', () => {
-    const wrapper = shallow(<BidCycle {...props} />);
+    const wrapper = shallow(<BidCycleCard {...props} />);
     expect(toJSON(wrapper)).toMatchSnapshot();
   });
 });
