+++ src/Components/BidCycleList/BidCycleList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { find } from 'lodash';
-import BidCycle from './BidCycle';
+import BidCycleCard from './BidCycleCard';
 import { Row, Column } from '../Layout';
 import ProfileSectionTitle from '../ProfileSectionTitle';
 import { BID_CYCLES } from '../../Constants/PropTypes';
@@ -8,23 +8,23 @@ import Menu from '../../Constants/Menu';
 
 const title = find(Menu[0].children, { route: '/profile/cycles/' }).text;
 
-const BidCycles = ({ cycles }) => (
+const BidCycleList = ({ cycles }) => (
   <Row fluid>
     <Column>
       <ProfileSectionTitle title={title} />
       {cycles.map(cycle => (
-        <BidCycle key={cycle.name} cycle={cycle} />
+        <BidCycleCard key={cycle.name} cycle={cycle} />
       ))}
     </Column>
   </Row>
   );
 
-BidCycles.propTypes = {
+BidCycleList.propTypes = {
   cycles: BID_CYCLES,
 };
 
-BidCycles.defaultProps = {
+BidCycleList.defaultProps = {
   cycles: [],
 };
 
-export default BidCycles;
+export default BidCycleList;
