+++ src/actions/bidCycles.js
@@ -2,7 +2,7 @@ import api from '../api';
 
 export const RECEIVE_BID_CYCLES = 'RECEIVE_BID_CYCLES';
 
-const recieveBidCycles = data => ({
+const receiveBidCycles = data => ({
   type: RECEIVE_BID_CYCLES,
   data,
 });
@@ -10,7 +10,7 @@ const recieveBidCycles = data => ({
 export const fetchBidCycles = () => dispatch => (
   api.get('/bidcycle/').then((response) => {
     const items = response.data.results || [];
-    dispatch(recieveBidCycles(items));
+    dispatch(receiveBidCycles(items));
   })
 );
